+++ tests/runtime-converters.test.cjs
@@ -1,19 +1,21 @@
 /**
- * OpenCode Agent Frontmatter Conversion Tests
+ * Runtime Converter Tests — OpenCode + Gemini
  *
- * Validates that convertClaudeToOpencodeFrontmatter correctly converts
- * agent frontmatter for OpenCode compatibility when isAgent: true.
+ * Tests for small runtime-specific conversion functions from install.js.
+ * Larger runtime test suites (Copilot, Codex, Antigravity) have their own files.
  *
- * Bug: Without isAgent flag, the function strips name: (agents need it),
- * keeps color:/skills:/tools: record (should strip), and doesn't add
- * model: inherit / mode: subagent (required by OpenCode agents).
+ * OpenCode: convertClaudeToOpencodeFrontmatter (agent + command modes)
+ * Gemini: convertClaudeToGeminiAgent (frontmatter + tool mapping + body escaping)
  */
 
 const { test, describe } = require('node:test');
 const assert = require('node:assert');
 
 process.env.GSD_TEST_MODE = '1';
-const { convertClaudeToOpencodeFrontmatter } = require('../bin/install.js');
+const {
+  convertClaudeToOpencodeFrontmatter,
+  convertClaudeToGeminiAgent,
+} = require('../bin/install.js');
 
 // Sample Claude agent frontmatter (matches actual GSD agent format)
 const SAMPLE_AGENT = `---
@@ -139,5 +141,43 @@ describe('OpenCode command conversion (isAgent: false, default)', () => {
     const result = convertClaudeToOpencodeFrontmatter(SAMPLE_COMMAND);
     const frontmatter = result.split('---')[1];
     assert.ok(frontmatter.includes('description:'), 'description should be kept');
+  });
+});
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Gemini CLI agent conversion (merged from gemini-config.test.cjs)
+// ─────────────────────────────────────────────────────────────────────────────
+
+describe('convertClaudeToGeminiAgent', () => {
+  test('drops unsupported skills frontmatter while keeping converted tools', () => {
+    const input = `---
+name: gsd-codebase-mapper
+description: Explores codebase and writes structured analysis documents.
+tools: Read, Bash, Grep, Glob, Write
+color: cyan
+skills:
+  - gsd-mapper-workflow
+---
+
+<role>
+Use \${PHASE} in shell examples.
+</role>`;
+
+    const result = convertClaudeToGeminiAgent(input);
+    const frontmatter = result.split('---')[1] || '';
+
+    assert.ok(frontmatter.includes('name: gsd-codebase-mapper'), 'keeps name');
+    assert.ok(frontmatter.includes('description: Explores codebase and writes structured analysis documents.'), 'keeps description');
+    assert.ok(frontmatter.includes('tools:'), 'adds Gemini tools array');
+    assert.ok(frontmatter.includes('  - read_file'), 'maps Read -> read_file');
+    assert.ok(frontmatter.includes('  - run_shell_command'), 'maps Bash -> run_shell_command');
+    assert.ok(frontmatter.includes('  - search_file_content'), 'maps Grep -> search_file_content');
+    assert.ok(frontmatter.includes('  - glob'), 'maps Glob -> glob');
+    assert.ok(frontmatter.includes('  - write_file'), 'maps Write -> write_file');
+    assert.ok(!frontmatter.includes('color:'), 'drops unsupported color field');
+    assert.ok(!frontmatter.includes('skills:'), 'drops unsupported skills field');
+    assert.ok(!frontmatter.includes('gsd-mapper-workflow'), 'drops skills list items');
+    assert.ok(result.includes('$PHASE'), 'escapes ${PHASE} shell variable for Gemini');
+    assert.ok(!result.includes('${PHASE}'), 'removes Gemini template-string pattern');
   });
 });
